fix: count_pct支持非数值列
count_pct只做行数计数，不需要数值转换。之前对临时列也做astype(float)
导致字符串列(如城市/频段)抛ValueError。

修复：临时列直接复制原始列，与数据类型无关。
- 校验层已允许count_pct用于非数值列(错误提示也说可用)
- 执行层现在与校验层行为一致
- 字符串列count_pct测试通过: 北京0.5/上海0.3333/广州0.1667

+++ app/src/pivot_analyzer.py
@@ -936,8 +936,9 @@ def _group_aggregate(df, group_cols, value_cols, agg_funcs, task):
         df = df.copy()
         for tmp_col, orig_col in pct_tmp_cols.items():
             df[tmp_col] = df[orig_col].astype(float)
+        # count_pct 只做行数计数，与数据类型无关，直接复制原始列（支持字符串列）
         for tmp_col, orig_col in count_pct_tmp_cols.items():
-            df[tmp_col] = df[orig_col].astype(float)
+            df[tmp_col] = df[orig_col]
     
     for vcol, funcs in field_funcs.items():
         agg_dict[vcol] = funcs
